pyproject.toml
[project]
name = "aktt-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot[job-queue]>=22.6",
    "supabase>=2.28.0",
]
